+++ src/ui/index.ts
@@ -74,6 +74,7 @@ export type {
   ResidentSelectorProps,
   ResidentSelectorMode,
   ResidentSelectorResident,
+  ResidentSelectorCreateResident,
   ResidentSelectorCreateData,
   ResidentSelectorSearchData,
   ResidentSelectorTab,